+++ docker-entrypoint.sh
@@ -1,12 +1,21 @@
 #!/bin/bash
 
+if [ -z "$APPLICATION_ENV" ]; then
+    export APPLICATION_ENV=production
+fi
+echo "APPLICATION_ENV=$APPLICATION_ENV"
+
 case "$1" in
     bot)
         echo "Starting discord bot..."
-        exec ruby -Ilib bin/lerk
+        exec ruby bin/lerk
+        ;;
+    migrate)
+        echo "Applying database migrations"
+        exec rake db:migrate
         ;;
     *)
         echo "Don't know what to do with $1"
-        echo "Valid commands: bot"
+        echo "Valid commands: bot, migrate"
         exit 1
 esac