+++ lib/lerk.rb
@@ -1,10 +1,13 @@
 # coding: utf-8
 
-require 'lerk/config'
+require 'lerk/models' unless ENV['LERK_SKIP_MODELS'] == '1'
 
 require 'lerk/lerk'
 require 'lerk/version'
 require 'lerk/excuse'
+require 'lerk/statistics'
+
+require 'lerk/util'
 
 require 'lerk/worker'