+++ lib/lerk/models.rb
@@ -1,0 +1,3 @@
+require 'lerk/models/discord_user'
+require 'lerk/models/event'
+require 'lerk/models/event_counter'